requirements.txt
numpy>=1.20.0
pandas>=1.3.0
scipy>=1.7.0
matplotlib>=3.4.0
openpyxl>=3.0.0
pyautogui>=0.9.53
keyboard>=0.13.5
pillow>=9.0.0
rainbow-api>=1.0.0
pybaselines>=1.0.0
lmfit>=1.2.0
